sports_marketing/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image_and_video"
version = "0.1.0"
description = "Image and video processing application"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Project Author"}
]
dependencies = [
    "streamlit>=1.22.0",
    "Pillow>=9.0.0",
    "boto3>=1.26.0",
    "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "checkov",
    "pre-commit",
    "black",
    "ruff",
    "pytest",
    "pytest-mock",
    "pytest-cov",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["image_and_video"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "I"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "--cov=image_and_video"

[tool.coverage.run]
source = ["image_and_video"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
